Implemented SDL, successful reflection, and openMP. Still some small error. Probably in calculating some of the primary ray attributes such as L or V given that the lights on each object are slightly tilted.

+++ RayTracer/Light.h
@@ -7,8 +7,8 @@
 class Light{
 private:
 public:
-	glm::vec4 pos;
-	glm::vec3 color;
+	glm::dvec4 pos;
+	glm::dvec3 color;
 
 };
