+++ RayTracer/Object.h
@@ -1,6 +1,7 @@
 #ifndef OBJECT_H
 #define OBJECT_H
 
+
 #include <iostream>
 #include <glm/glm.hpp>
 #include "Intersection.h"
@@ -8,14 +9,14 @@
 class Object{
 private:
 public:
-	glm::vec3 md = glm::vec3(1, .1, .1);
-	glm::vec3 ms = glm::vec3(1, .1, .1);
-	float shininess = 60.0;
+	glm::dvec3 md = glm::dvec3(1, .1, .1);
+	glm::dvec3 ms = glm::dvec3(1, .1, .1);
+	double shininess = 60.0;
 
 
-	glm::mat4 xForm;
-	glm::mat4 ixForm;
-	glm::mat4 ixtForm;
+	glm::dmat4 xForm;
+	glm::dmat4 ixForm;
+	glm::dmat4 ixtForm;
 
 	virtual Intersection intersect(Ray ray){ return Intersection(); };
 };
