+++ RayTracer/Viewport.h
@@ -9,14 +9,14 @@
 
 class Viewport{
 private:
-	std::vector<glm::vec3> pixels;
+	std::vector<glm::dvec3> pixels;
 
 public:
-	float d = .5;
+	double d = .5;
 	int width = 1000, height = 1000;
 	int clipDistance = 10000000;
 
-	glm::vec3 backColor = glm::vec3(0.0f, 0.0f, 1.0f);
+	glm::dvec3 backColor = glm::dvec3(0.0, 0.0, 1.0);
 
 	Viewport(){
 		pixels.resize(width * height);
@@ -24,11 +24,11 @@ public:
 
 	void recalculate(){ pixels.resize(width*height); }
 
-	glm::vec3 getPixel(int x, int y){
+	glm::dvec3 getPixel(int x, int y){
 		return pixels[x*width + y];
 	}
 
-	void setPixel(int x, int y, glm::vec3 input){
+	void setPixel(int x, int y, glm::dvec3 input){
 		pixels[x*width + y] = input;
 	}
 };
